Add files via upload
Changed GUI features by adding in a minimum and maximum time interval for the clicker. The default is 2 - 3 seconds. Also added bindings to prevent text field from being anything other then an integer value. This will need to be updated in a future version so that double values can be used.